#418 - run importer v1. Next up, various mirror models via handles

+++ OneSila/sales_channels/integrations/woocommerce/tests/tests_factories/mixins.py
@@ -11,6 +11,8 @@ logger = logging.getLogger(__name__)
 
 
 class TestCaseWoocommerceMixin(TransactionTestCase):
+    remove_woocommerce_mirror_and_remote_on_teardown = True
+
     def setUp(self):
         """
         Set up the test case with a mock WoocommerceSalesChannel.
@@ -52,19 +54,20 @@ class TestCaseWoocommerceMixin(TransactionTestCase):
         remote_currency.save()
 
     def tearDown(self):
-        try:
-            from sales_channels.integrations.woocommerce.models import WoocommerceGlobalAttribute, \
-                WoocommerceProduct
+        if self.remove_woocommerce_mirror_and_remote_on_teardown:
+            try:
+                from sales_channels.integrations.woocommerce.models import WoocommerceGlobalAttribute, \
+                    WoocommerceProduct
 
-            attribute_ids = WoocommerceGlobalAttribute.objects.filter(sales_channel=self.sales_channel).values_list('remote_id', flat=True)
-            product_ids = WoocommerceProduct.objects.filter(sales_channel=self.sales_channel).values_list('remote_id', flat=True)
+                attribute_ids = WoocommerceGlobalAttribute.objects.filter(sales_channel=self.sales_channel).values_list('remote_id', flat=True)
+                product_ids = WoocommerceProduct.objects.filter(sales_channel=self.sales_channel).values_list('remote_id', flat=True)
 
-            for attribute_id in attribute_ids:
-                self.api.delete_attribute(attribute_id)
+                for attribute_id in attribute_ids:
+                    self.api.delete_attribute(attribute_id)
 
-            for product_id in product_ids:
-                self.api.delete_product(product_id)
+                for product_id in product_ids:
+                    self.api.delete_product(product_id)
 
-        except AttributeError:
-            logger.warning("No api set. Cleanup is not possible.")
-            pass
+            except AttributeError:
+                logger.warning("No api set. Cleanup is not possible.")
+                pass
